+++ mzLib/Test/Omics/Modifications/ModificationEqualityTests.cs
@@ -1,7 +1,7 @@
-﻿// Copyright 2012, 2013, 2014 Derek J. Bailey
+// Copyright 2012, 2013, 2014 Derek J. Bailey
 // Modified work copyright 2016 Stefan Solntsev
 //
-// This file (TestFragments.cs) is part of Proteomics.
+// This file (ModificationEqualityTests.cs) is part of Proteomics.
 //
 // Proteomics is free software: you can redistribute it and/or modify it
 // under the terms of the GNU Lesser General Public License as published
@@ -39,7 +39,7 @@ namespace Test
 {
     [TestFixture]
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    public sealed class TestModifications
+    public static class ModificationEqualityTests
     {
         private static Stopwatch Stopwatch { get; set; }
 
@@ -86,7 +86,7 @@ namespace Test
         }
 
         [Test]
-        public void NameAndSites()
+        public static void NameAndSites()
         {
             // Empty modification, has no name and by default has Sites = ModificationSites.Any
             OldSchoolModification a = new OldSchoolModification();
@@ -95,7 +95,7 @@ namespace Test
         }
 
         [Test]
-        public void ModificationEquality()
+        public static void ModificationEquality()
         {
             // Empty modification, has no name and by default has Sites = ModificationSites.Any
             OldSchoolModification a = new OldSchoolModification();
@@ -108,7 +108,7 @@ namespace Test
         }
 
         [Test]
-        public void ModificationSitesTest()
+        public static void ModificationSitesTest()
         {
             // Empty modification, has no name and by default has Sites = ModificationSites.Any
             var a = ModificationSites.A | ModificationSites.E;
@@ -116,7 +116,7 @@ namespace Test
         }
 
         [Test]
-        public void Sites()
+        public static void Sites()
         {
             // Empty modification, has no name and by default has Sites = ModificationSites.Any
             var a = ModificationSites.A | ModificationSites.C | ModificationSites.E;
@@ -130,7 +130,7 @@ namespace Test
         }
 
         [Test]
-        public void SilacLabelTest()
+        public static void SilacLabelTest()
         {
             Protein originalProtein = new Protein("ACDEFGHIKAKAK", "TEST");
             List<PeptideWithSetModifications> originalDigest = originalProtein.Digest(new DigestionParams(), new List<Modification>(), new List<Modification>()).ToList();
@@ -208,10 +208,8 @@ namespace Test
             Assert.IsTrue(silacDigest.Count == 14);
         }
 
-
-
         [Test]
-        public void ModificationCollectionTest()
+        public static void ModificationCollectionTest()
         {
             ModificationCollection a = new ModificationCollection(new OldSchoolModification(1, "Mod1"), new OldSchoolModification(2, "Mod2"));
 
@@ -242,14 +240,14 @@ namespace Test
         }
 
         [Test]
-        public void ModificationCollectionTest2()
+        public static void ModificationCollectionTest2()
         {
             ModificationCollection a = new ModificationCollection(new OldSchoolModification(1, "Mod1"), new OldSchoolModification(2, "Mod2"));
             Assert.IsFalse(a.Remove(new OldSchoolModification(3, "Mod3")));
         }
 
         [Test]
-        public void ModificationWithMultiplePossibilitiesTest()
+        public static void ModificationWithMultiplePossibilitiesTest()
         {
             var m = new ModificationWithMultiplePossibilitiesCollection("My Iso Mod", ModificationSites.E);
             m.AddModification(new OldSchoolModification(1, "My Mod1a", ModificationSites.E));
@@ -266,7 +264,7 @@ namespace Test
         }
 
         [Test]
-        public void ModificationCollectionTestTest()
+        public static void ModificationCollectionTestTest()
         {
             OldSchoolModification mod1 = new OldSchoolModification(10, "mass 10 modification");
             OldSchoolModification mod2 = new OldSchoolModification(100, "mass 100 modification");
@@ -279,7 +277,7 @@ namespace Test
         }
 
         [Test]
-        public void ModificationSitesTest55()
+        public static void ModificationSitesTest55()
         {
             Assert.IsTrue(ModificationSites.E.ContainsSites(ModificationSites.Any));
             Assert.IsFalse(ModificationSites.E.ContainsSites(ModificationSites.None));
@@ -287,7 +285,7 @@ namespace Test
         }
 
         [Test]
-        public void ChemicalFormulaModification()
+        public static void ChemicalFormulaModification()
         {
             OldSchoolChemicalFormulaModification a = new OldSchoolChemicalFormulaModification(ChemicalFormula.ParseFormula("OH"));
             OldSchoolChemicalFormulaModification b = new OldSchoolChemicalFormulaModification(a);
@@ -295,7 +293,7 @@ namespace Test
         }
 
         [Test]
-        public void ModificationCollectionScrambledEquals()
+        public static void ModificationCollectionScrambledEquals()
         {
             ModificationCollection a = new ModificationCollection(new OldSchoolModification(1, "Mod1"), new OldSchoolModification(2, "Mod2"));
             ModificationCollection b = new ModificationCollection(new OldSchoolModification(1, "Mod1"), new OldSchoolModification(3, "Mod3"));
@@ -304,7 +302,7 @@ namespace Test
         }
 
         [Test]
-        public void Test_modification_hash_set()
+        public static void Test_modification_hash_set()
         {
             ModificationMotif.TryGetMotif("K", out ModificationMotif motif);
             Modification m1 = new Modification("23", null, "unknown", null, motif, null, null, 42.01, null, null, null, null, null, null);
@@ -314,7 +312,7 @@ namespace Test
         }
 
         [Test]
-        public void Test_modification2_hash_set()
+        public static void Test_modification2_hash_set()
         {
             ModificationMotif.TryGetMotif("K", out ModificationMotif motif);
             Modification m1 = new Modification("id1", null, "modificationType", null, motif, "Anywhere.", null, 42.01, new Dictionary<string, IList<string>>(), null, null, null, null, null);
@@ -327,7 +325,7 @@ namespace Test
         }
 
         [Test]
-        public void Test_modification3_hash_set() // numerical tolerance is 1e-9 so these two mods need to evaluate as identical
+        public static void Test_modification3_hash_set() // numerical tolerance is 1e-9 so these two mods need to evaluate as identical
         {
             ModificationMotif.TryGetMotif("K", out ModificationMotif motif);
             Modification m1 = new Modification(_originalId: "id1", _modificationType: "modificationType", _target: motif, _locationRestriction: "Anywhere.", _monoisotopicMass: 1.11111d, _databaseReference: new Dictionary<string, IList<string>>(), _neutralLosses: new Dictionary<DissociationType, List<double>> { { DissociationType.AnyActivationType, new List<double> { 2.222222 } } }, _diagnosticIons: new Dictionary<DissociationType, List<double>> { { DissociationType.AnyActivationType, new List<double> { 1.2233 } } });
@@ -341,7 +339,7 @@ namespace Test
         }
 
         [Test]
-        public void TestInvalidModificationHash()
+        public static void TestInvalidModificationHash()
         {
             ModificationMotif.TryGetMotif("K", out ModificationMotif motif);
             Modification m1 = new Modification(_originalId: "id1", _modificationType: "modificationType", _target: motif, _locationRestriction: "Anywhere.");
@@ -369,388 +367,6 @@ namespace Test
             Assert.IsFalse(m2.ValidModification);
             Assert.False(m1.Equals(m2));
             Assert.AreEqual(2, mods.Count);
-        }
-
-        [Test]
-        public void TestFragmentationNoMod()
-        {
-            // First we're checking to see if the fragment masses of an unmodified peptide a calculated correctly
-            var prot = new Protein("PEPTIDE", null);
-            DigestionParams digestionParams = new DigestionParams(
-
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-            List<Modification> variableModifications = new List<Modification>();
-            var ye = prot.Digest(digestionParams, new List<Modification>(), variableModifications).ToList();
-
-            // check unmodified
-            var unmodPeptide = ye.Where(p => p.AllModsOneIsNterminus.Count == 0).First();
-            var fragments = new List<Product>();
-            unmodPeptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, fragments);
-            var myUnmodFragmentMasses = fragments.Select(v => (int)Math.Round(v.NeutralMass.ToMz(1), 1)).ToList();
-            HashSet<int> expectedMzs = new HashSet<int> { 98, 227, 324, 425, 538, 653, 703, 574, 477, 376, 263, 148 };
-
-            Assert.That(expectedMzs.SetEquals(myUnmodFragmentMasses));
-        }
-
-        [Test]
-        public void TestFragmentationModNoNeutralLoss()
-        {
-            // Now we'll check the mass of modified peptide with no neutral losses
-            ModificationMotif.TryGetMotif("T", out ModificationMotif motif);
-            Modification mod = new Modification(_originalId: "oxidation", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("O1"), _locationRestriction: "Anywhere.");
-            List<Modification> modlist = new List<Modification> { mod };
-            DigestionParams digestionParams = new DigestionParams(
-                protease: "trypsin",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            var prot = new Protein("PEPTIDE", null, oneBasedModifications: new Dictionary<int, List<Modification>> { { 4, modlist } });
-            var ye = prot.Digest(digestionParams, new List<Modification>(), new List<Modification>()).ToList();
-
-            // check unmodified
-            var unmodPeptide = ye.Where(p => p.AllModsOneIsNterminus.Count == 0).First();
-            var myUnmodFragments = new List<Product>();
-            unmodPeptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, myUnmodFragments);
-            var neutralMasses = new List<double>();
-            neutralMasses.AddRange(myUnmodFragments.Select(m => m.NeutralMass).ToList());
-            var expectedMasses = new List<double> { 97, 226, 323, 424, 537, 652, 147, 262, 375, 476, 573, 702 };
-            for (int i = 0; i < neutralMasses.Count; i++)
-            {
-                neutralMasses[i] = Chemistry.ClassExtensions.RoundedDouble(neutralMasses[i], 0).Value;
-            }
-
-            var firstNotSecond = neutralMasses.Except(expectedMasses).ToList();
-            var secondNotFirst = expectedMasses.Except(neutralMasses).ToList();
-
-            //this is the set without oxidation
-            Assert.AreEqual(12, myUnmodFragments.Count);
-            Assert.AreEqual(0, firstNotSecond.Count);
-            Assert.AreEqual(0, secondNotFirst.Count);
-
-            // with oxidation, no neutral loss
-            var modPeptide = ye.Where(p => p.AllModsOneIsNterminus.Count == 1).First();
-
-            var myModFragments = new List<Product>();
-            modPeptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, myModFragments);
-            neutralMasses = new List<double>();
-            neutralMasses.AddRange(myModFragments.Select(m => m.NeutralMass).ToList());
-            expectedMasses = new List<double> { 97, 226, 323, 440, 553, 668, 147, 262, 375, 492, 589, 718 };
-            for (int i = 0; i < neutralMasses.Count; i++)
-            {
-                neutralMasses[i] = Chemistry.ClassExtensions.RoundedDouble(neutralMasses[i], 0).Value;
-            }
-
-            firstNotSecond = neutralMasses.Except(expectedMasses).ToList();
-            secondNotFirst = expectedMasses.Except(neutralMasses).ToList();
-
-            //this is the set with oxidation
-            Assert.AreEqual(12, myUnmodFragments.Count);
-            Assert.AreEqual(0, firstNotSecond.Count);
-            Assert.AreEqual(0, secondNotFirst.Count);
-        }
-
-        [Test]
-        public void Test_FragmentationModNeutralLoss()
-        {
-            // Now we'll check the mass of modified peptide with no neutral losses
-            ModificationMotif.TryGetMotif("T", out ModificationMotif motif);
-            Modification mod = new Modification(_originalId: "phospho", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("H1 O3 P1"), _neutralLosses: new Dictionary<DissociationType, List<double>> { { DissociationType.HCD, new List<double> { ChemicalFormula.ParseFormula("H3 O4 P1").MonoisotopicMass } } }, _locationRestriction: "Anywhere.");
-            List<Modification> modlist = new List<Modification> { mod };
-            DigestionParams digestionParams = new DigestionParams(
-                protease: "trypsin",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            var prot = new Protein("PEPTIDE", null, oneBasedModifications: new Dictionary<int, List<Modification>> { { 4, modlist } });
-            var ye = prot.Digest(digestionParams, new List<Modification>(), new List<Modification>()).ToList();
-
-            var peptideWithNeutralMassMod = ye.Where(v => v.AllModsOneIsNterminus.Count > 0).First();
-
-            var myModFragments = new List<Product>();
-            peptideWithNeutralMassMod.Fragment(DissociationType.HCD, FragmentationTerminus.Both, myModFragments);
-            HashSet<int> neutralMasses = new HashSet<int>(myModFragments.Select(m => (int)m.NeutralMass.ToMz(1)).ToList());
-            HashSet<int> expectedMasses = new HashSet<int> { 98,227, 324, 407, 520, 635, 505, 618, 733, //b-ions with and without neutral loss
-            148, 263, 376, 459, 556, 685, 557, 654, 783, //y-ions with and without neutral loss
-            782}; //molecular ion with neutral loss
-
-            CollectionAssert.AreEquivalent(neutralMasses, expectedMasses);
-        }
-
-        [Test]
-        public void Test_FragmentationTwoModNeutralLoss()
-        {
-            // Now we'll check the mass of modified peptide with 2 neutral loss mods
-            ModificationMotif.TryGetMotif("Q", out ModificationMotif motifone);
-            Modification modone = new Modification(_originalId: "ammonia", _modificationType: "testModType", _target: motifone, _monoisotopicMass: 0, _neutralLosses: new Dictionary<DissociationType,
-                List<double>> { { DissociationType.HCD, new List<double> { ChemicalFormula.ParseFormula("H3 N1").MonoisotopicMass } } }, _locationRestriction: "Anywhere.");
-
-            ModificationMotif.TryGetMotif("T", out ModificationMotif motiftwo);
-            Modification modtwo = new Modification(_originalId: "phospho", _modificationType: "testModType", _target: motiftwo, _chemicalFormula: ChemicalFormula.ParseFormula("H1 O3 P1"), _neutralLosses: new Dictionary<DissociationType,
-                List<double>> { { DissociationType.HCD, new List<double> { ChemicalFormula.ParseFormula("H3 O4 P1").MonoisotopicMass } } }, _locationRestriction: "Anywhere.");
-
-            List<Modification> modlistone = new List<Modification> { modone };
-            List<Modification> modlisttwo = new List<Modification> { modtwo };
-
-            DigestionParams digestionParams = new DigestionParams(
-                protease: "trypsin",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            var prot = new Protein("PEQTIDE", null, oneBasedModifications: new Dictionary<int, List<Modification>> { { 3, modlistone }, { 4, modlisttwo } });
-            var ye = prot.Digest(digestionParams, new List<Modification>(), new List<Modification>()).ToList();
-
-            var peptideWithNeutralMassMod = ye.Where(v => v.AllModsOneIsNterminus.Count == 2).First();
-
-            var myModFragments = new List<Product>();
-            peptideWithNeutralMassMod.Fragment(DissociationType.HCD, FragmentationTerminus.Both, myModFragments);
-            HashSet<int> neutralMasses = new HashSet<int>(myModFragments.Select(m => (int)m.NeutralMass.ToMz(1)).ToList());
-            HashSet<int> expectedMasses = new HashSet<int> { 98, 227, 355, 536, 649, 764, 438, 551, 666, 338, 519, 632, 747, // b-ions with and without neutral losses
-                                                             148, 263, 376, 557, 685, 814, 668, 797, 459, 587, 716, //y ions with and without neutral losses
-                                                               813, 894, }; //molecular ion with neutral losses (phospho and ammonia respectively)
-
-            CollectionAssert.AreEquivalent(neutralMasses, expectedMasses);
-        }
-
-        [Test]
-        public void Test_FragmentationTwoModNeutralLossTwoFragTypes()
-        {
-            // Now we'll check the mass of modified peptide with no neutral losses
-            ModificationMotif.TryGetMotif("T", out ModificationMotif motif);
-
-            Dictionary<DissociationType, List<double>> myNeutralLosses = new Dictionary<DissociationType, List<double>>()
-            {
-                { DissociationType.HCD, new List<double> { ChemicalFormula.ParseFormula("H3 O4 P1").MonoisotopicMass } },
-                { DissociationType.ETD, new List<double>() { ChemicalFormula.ParseFormula("H3 N1").MonoisotopicMass } } // this makes no sense in real life, it's just for a unit test
-            };
-
-            Modification mod = new Modification(_originalId: "phospho", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("H1 O3 P1"), _neutralLosses: myNeutralLosses, _locationRestriction: "Anywhere.");
-            List<Modification> modlist = new List<Modification> { mod };
-            DigestionParams digestionParams = new DigestionParams(
-                protease: "trypsin",
-                maxMissedCleavages: 0,
-                minPeptideLength: 1,
-                initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            var prot = new Protein("PEPTIDE", null, oneBasedModifications: new Dictionary<int, List<Modification>> { { 4, modlist } });
-            var ye = prot.Digest(digestionParams, new List<Modification>(), new List<Modification>()).ToList();
-
-            var peptideWithNeutralMassMod = ye.Where(v => v.AllModsOneIsNterminus.Count == 1).First();
-
-            var myModFragmentsHCD = new List<Product>();
-            peptideWithNeutralMassMod.Fragment(DissociationType.HCD, FragmentationTerminus.Both, myModFragmentsHCD);
-
-            var neutralMassesHCD = myModFragmentsHCD.Select(m => (int)m.NeutralMass.ToMz(1));
-            var expectedMassesHCD = new HashSet<int> { 98, 227, 324, 407, 520, 635, 505, 618, 733,// b-ions with and without neutral loss
-                                                        148, 263, 376, 459, 556, 685, 557, 654, 783,//y-ions with and without neutral loss
-                                                        782};// molecular ion with neutral loss
-
-            CollectionAssert.AreEquivalent(expectedMassesHCD, neutralMassesHCD);
-
-            //Now try the other half
-            var myModFragmentsETD = new List<Product>();
-            peptideWithNeutralMassMod.Fragment(DissociationType.ETD, FragmentationTerminus.Both, myModFragmentsETD);
-
-            var neutralMassesETD = myModFragmentsETD.Select(m => (int)m.NeutralMass.ToMz(1));
-            var expectedMassesETD = new HashSet<int> { 115, 244, 341, 505, 618, 733, 522, 635, 750,  // c-ions and c-17 ions
-            148, 263, 376, 540, 637, 766, 557, 654, 783,          // y and y-17 ions
-            133, 248, 361, 525, 622, 751, 542, 639, 768,         // z+1 and z+1-17 ions
-            863 };//Molecular ions minus ammonia
-
-            CollectionAssert.AreEquivalent(expectedMassesHCD, neutralMassesHCD);
-        }
-        
-        [Test]
-        public static void TestSerializationPeptideFromString()
-        {
-            // purpose of this test is to serialize/deserialize a PeptideWithSetModifications and make sure the deserialized peptide
-            // has the same properties as before it was serialized. This peptide is unmodified and generated from reading in a string
-            string sequence = "PEPTIDE";
-            PeptideWithSetModifications peptide = new PeptideWithSetModifications(sequence, new Dictionary<string, Modification>(), 0, null, null, 1, 7, 0);
-            PeptideWithSetModifications deserializedPeptide = null;
-
-            string dir = System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, "TestSerializationPeptideFromString");
-            System.IO.Directory.CreateDirectory(dir);
-            string path = System.IO.Path.Combine(dir, "myPeptideIndex.ind");
-
-            var messageTypes = typeof(PeptideWithSetModifications);
-            var ser = new NetSerializer.Serializer(new List<Type> { messageTypes });
-
-            using (var file = System.IO.File.Create(path))
-            {
-                ser.Serialize(file, peptide);
-            }
-
-            using (var file = System.IO.File.OpenRead(path))
-            {
-                deserializedPeptide = (PeptideWithSetModifications)ser.Deserialize(file);
-            }
-
-            deserializedPeptide.SetNonSerializedPeptideInfo(new Dictionary<string, Modification>(), new Dictionary<string, Protein>(), null);
-
-            // not asserting any protein properties - since the peptide was created from a sequence string it didn't have a protein to begin with
-
-            Assert.That(peptide.Equals(deserializedPeptide));
-            Assert.That(deserializedPeptide.MonoisotopicMass == peptide.MonoisotopicMass);
-            Assert.That(deserializedPeptide.SequenceWithChemicalFormulas == peptide.SequenceWithChemicalFormulas);
-
-            var products = new List<Product>();
-
-            deserializedPeptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, products);
-            List<double> deserializedPeptideFragments = products.Select(v => v.NeutralMass).ToList();
-
-            peptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, products);
-            List<double> peptideFragments = products.Select(v => v.NeutralMass).ToList();
-
-            Assert.That(deserializedPeptideFragments.SequenceEqual(peptideFragments));
-        }
-
-        [Test]
-        public static void TestSerializationPeptideFromProtein()
-        {
-            // purpose of this test is to serialize/deserialize a PeptideWithSetModifications and make sure the deserialized peptide
-            // has the same properties as before it was serialized. This peptide is unmodified and generated from digesting a protein
-            Protein protein = new Protein("PEPTIDE", "Accession1", name: "MyProtein");
-
-            PeptideWithSetModifications peptide = protein.Digest(new DigestionParams(), new List<Modification>(), new List<Modification>()).First();
-            PeptideWithSetModifications deserializedPeptide = null;
-
-            string dir = System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, "TestSerializationPeptideFromProtein");
-            System.IO.Directory.CreateDirectory(dir);
-            string path = System.IO.Path.Combine(dir, "myPeptideIndex.ind");
-
-            var messageTypes = typeof(PeptideWithSetModifications);
-            var ser = new NetSerializer.Serializer(new List<Type> { messageTypes });
-
-            using (var file = System.IO.File.Create(path))
-            {
-                ser.Serialize(file, peptide);
-            }
-
-            using (var file = System.IO.File.OpenRead(path))
-            {
-                deserializedPeptide = (PeptideWithSetModifications)ser.Deserialize(file);
-            }
-
-            deserializedPeptide.SetNonSerializedPeptideInfo(new Dictionary<string, Modification>(), new Dictionary<string, Protein> { { protein.Accession, protein } }, peptide.DigestionParams);
-
-            Assert.That(peptide.DigestionParams.Equals(deserializedPeptide.DigestionParams));
-            Assert.That(peptide.Equals(deserializedPeptide));
-            Assert.That(deserializedPeptide.Protein.Name == peptide.Protein.Name);
-            Assert.That(deserializedPeptide.MonoisotopicMass == peptide.MonoisotopicMass);
-            Assert.That(deserializedPeptide.SequenceWithChemicalFormulas == peptide.SequenceWithChemicalFormulas);
-
-            var products = new List<Product>();
-
-            deserializedPeptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, products);
-            List<double> deserializedPeptideFragments = products.Select(v => v.NeutralMass).ToList();
-
-            peptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, products);
-            List<double> peptideFragments = products.Select(v => v.NeutralMass).ToList();
-
-            Assert.That(deserializedPeptideFragments.SequenceEqual(peptideFragments));
-        }
-
-        [Test]
-        public static void TestSerializationPeptideFromProteinWithMod()
-        {
-            // purpose of this test is to serialize/deserialize a PeptideWithSetModifications and make sure the deserialized peptide
-            // has the same properties as before it was serialized. This peptide is modified with a phosphorylation
-
-            ModificationMotif.TryGetMotif("T", out ModificationMotif motif);
-
-            Dictionary<DissociationType, List<double>> myNeutralLosses = new Dictionary<DissociationType, List<double>>()
-            {
-                { DissociationType.HCD, new List<double> { ChemicalFormula.ParseFormula("H3 O4 P1").MonoisotopicMass } },
-                { DissociationType.ETD, new List<double>() { ChemicalFormula.ParseFormula("H3 N1").MonoisotopicMass } } // this makes no sense in real life, it's just for a unit test
-            };
-
-            Modification mod = new Modification(_originalId: "phospho", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("H1 O3 P1"), _neutralLosses: myNeutralLosses, _locationRestriction: "Anywhere.");
-
-            Dictionary<int, List<Modification>> mods = new Dictionary<int, List<Modification>> { { 4, new List<Modification> { mod } } };
-
-            Protein protein = new Protein("PEPTIDE", "Accession1", name: "MyProtein", oneBasedModifications: mods);
-
-            PeptideWithSetModifications peptide = protein.Digest(new DigestionParams(), new List<Modification>(), new List<Modification>()).Where(v => v.AllModsOneIsNterminus.Count == 1).First();
-            PeptideWithSetModifications deserializedPeptide = null;
-
-            string dir = System.IO.Path.Combine(TestContext.CurrentContext.TestDirectory, "TestSerializationPeptideFromProteinWithMod");
-            System.IO.Directory.CreateDirectory(dir);
-            string path = System.IO.Path.Combine(dir, "myPeptideIndex.ind");
-
-            var messageTypes = typeof(PeptideWithSetModifications);
-            var ser = new NetSerializer.Serializer(new List<Type> { messageTypes });
-
-            using (var file = System.IO.File.Create(path))
-            {
-                ser.Serialize(file, peptide);
-            }
-
-            using (var file = System.IO.File.OpenRead(path))
-            {
-                deserializedPeptide = (PeptideWithSetModifications)ser.Deserialize(file);
-            }
-
-            Dictionary<string, Modification> stringToMod = new Dictionary<string, Modification> { { mods.Values.First().First().IdWithMotif, mods.Values.First().First() } };
-
-            deserializedPeptide.SetNonSerializedPeptideInfo(stringToMod, new Dictionary<string, Protein> { { protein.Accession, protein } }, peptide.DigestionParams);
-
-            Assert.That(peptide.Equals(deserializedPeptide));
-            Assert.That(deserializedPeptide.Protein.Name == peptide.Protein.Name);
-            Assert.That(deserializedPeptide.MonoisotopicMass == peptide.MonoisotopicMass);
-            Assert.That(deserializedPeptide.SequenceWithChemicalFormulas == peptide.SequenceWithChemicalFormulas);
-
-            var products = new List<Product>();
-
-            deserializedPeptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, products);
-            List<double> deserializedPeptideFragments = products.Select(v => v.NeutralMass).ToList();
-
-            peptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, products);
-            List<double> peptideFragments = products.Select(v => v.NeutralMass).ToList();
-
-            Assert.That(deserializedPeptideFragments.SequenceEqual(peptideFragments));
-        }
-
-        [Test]
-        public static void TestFragmentNterminalModifiedPeptide()
-        {
-            ModificationMotif.TryGetMotif("P", out ModificationMotif motif);
-            Modification nTermMod = new Modification(_originalId: "acetylation", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("C2H2O1"), _locationRestriction: "N-terminal.");
-
-            Dictionary<int, List<Modification>> mods = new Dictionary<int, List<Modification>> { { 1, new List<Modification> { nTermMod } } };
-
-            Protein protein = new Protein("PEPTIDE", "", oneBasedModifications: mods);
-            PeptideWithSetModifications peptide = protein.Digest(new DigestionParams(), new List<Modification>(), new List<Modification>()).Where(p => p.AllModsOneIsNterminus.Count == 1).First();
-            Assert.That(peptide.FullSequence == "[testModType:acetylation on P]PEPTIDE");
-
-            var fragments = new List<Product>();
-            peptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, fragments);
-
-            var roundedFragments = fragments.Select(f => (int)f.NeutralMass).ToList();
-            CollectionAssert.AreEquivalent(roundedFragments, new int[] { 139, 268, 365, 466, 579, 694, 147, 262, 375, 476, 573, 702 });
-        }
-
-        [Test]
-        public static void TestFragmentCTerminalModifiedPeptide()
-        {
-            ModificationMotif.TryGetMotif("E", out ModificationMotif motif);
-            Modification cTermMod = new Modification(_originalId: "acetylation", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("C2H2O1"), _locationRestriction: "C-terminal.");
-
-            Dictionary<int, List<Modification>> mods = new Dictionary<int, List<Modification>> { { 7, new List<Modification> { cTermMod } } };
-
-            Protein protein = new Protein("PEPTIDE", "", oneBasedModifications: mods);
-            PeptideWithSetModifications peptide = protein.Digest(new DigestionParams(), new List<Modification>(), new List<Modification>()).Where(p => p.AllModsOneIsNterminus.Count == 1).First();
-            Assert.That(peptide.FullSequence == "PEPTIDE-[testModType:acetylation on E]");
-
-            var fragments = new List<Product>();
-            peptide.Fragment(DissociationType.HCD, FragmentationTerminus.Both, fragments);
-
-            var roundedFragments = fragments.Select(f => (int)f.NeutralMass).ToList();
-            CollectionAssert.AreEquivalent(roundedFragments, new int[] { 97, 226, 323, 424, 537, 652, 189, 304, 417, 518, 615, 744 });
         }
 
         [Test]
@@ -805,7 +421,7 @@ namespace Test
         }
 
         [Test]
-        public void CompareTo_SameModification_ReturnsZero()
+        public static void CompareTo_SameModification_ReturnsZero()
         {
             ModificationMotif.TryGetMotif("A", out var motif);
             var chemicalFormula = new ChemicalFormula();
@@ -816,7 +432,7 @@ namespace Test
         }
 
         [Test]
-        public void CompareTo_DifferentIdWithMotif_ReturnsNonZero()
+        public static void CompareTo_DifferentIdWithMotif_ReturnsNonZero()
         {
             ModificationMotif.TryGetMotif("A", out var motif);
             var chemicalFormula = new ChemicalFormula();
@@ -828,7 +444,7 @@ namespace Test
         }
 
         [Test]
-        public void CompareTo_DifferentModificationType_ReturnsNonZero()
+        public static void CompareTo_DifferentModificationType_ReturnsNonZero()
         {
             ModificationMotif.TryGetMotif("A", out var motif);
             var chemicalFormula = new ChemicalFormula();
@@ -840,7 +456,7 @@ namespace Test
         }
 
         [Test]
-        public void CompareTo_DifferentTarget_ReturnsNonZero()
+        public static void CompareTo_DifferentTarget_ReturnsNonZero()
         {
             ModificationMotif.TryGetMotif("A", out var motif1);
             ModificationMotif.TryGetMotif("B", out var motif2);
@@ -853,7 +469,7 @@ namespace Test
         }
 
         [Test]
-        public void CompareTo_DifferentLocationRestriction_ReturnsNonZero()
+        public static void CompareTo_DifferentLocationRestriction_ReturnsNonZero()
         {
             ModificationMotif.TryGetMotif("A", out var motif);
             var chemicalFormula = new ChemicalFormula();
@@ -865,7 +481,7 @@ namespace Test
         }
 
         [Test]
-        public void CompareTo_DifferentMonoisotopicMass_ReturnsNonZero()
+        public static void CompareTo_DifferentMonoisotopicMass_ReturnsNonZero()
         {
             ModificationMotif.TryGetMotif("A", out var motif);
             var chemicalFormula = new ChemicalFormula();
